now uses lm_eval cli instead
